+++ core/src/files.js
@@ -8,6 +8,8 @@ export * from './init.js';
 export * from './projects.js';
 export * from './guidance.js';
 export * from './memory.js';
+export * from './artifacts.js';
 export * from './schema.js';
+export { ulid } from './ids.js';
 export { slugify } from './slug.js';
 export { readDoc, writeDoc } from './frontmatter.js';